Move `describe()` and `from_json()` to the `evaluator` module
