+++ src/compiler/include/sourcemeta/blaze/compiler_output.h
@@ -255,20 +255,6 @@ private:
 };
 
 /// @ingroup compiler
-///
-/// This function translates a "post" step execution into a human-readable
-/// string. Useful as the building block for producing user-friendly evaluation
-/// results.
-///
-/// Note that describing a "pre" step execution is NOT supported.
-auto SOURCEMETA_BLAZE_COMPILER_EXPORT
-describe(const bool valid, const Instruction &step,
-         const sourcemeta::core::WeakPointer &evaluate_path,
-         const sourcemeta::core::WeakPointer &instance_location,
-         const sourcemeta::core::JSON &instance,
-         const sourcemeta::core::JSON &annotation) -> std::string;
-
-/// @ingroup compiler
 /// Represents standard output formats
 /// See
 /// https://json-schema.org/draft/2020-12/json-schema-core#name-output-structure
